v2d phaseCalInt type changed from 'int' to 'float' to match VEX, and extend the printed digits to support the entry of the 77.777 MHz spacing used by KVN-Sejong optical PCal system
git-svn-id: https://svn.atnf.csiro.au/difx@10966 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/vex2difx/trunk/vexdatamodel/vex_mode.cpp
@@ -447,7 +447,7 @@ void VexMode::setSampling(const std::string &antName, unsigned int streamId, enu
 	}
 }
 
-void VexMode::setPhaseCalInterval(const std::string &antName, int phaseCalIntervalMHz)
+void VexMode::setPhaseCalInterval(const std::string &antName, float phaseCalIntervalMHz)
 {
 	std::map<std::string,VexSetup>::iterator it = setups.find(antName);
 	if(it != setups.end())
